add swagger dependency

+++ src/main/java/com/example/demo/Services/TaskService.java
@@ -6,6 +6,8 @@ import com.example.demo.entities.Task;
 import com.example.demo.entities.User;
 import com.example.demo.repositories.TaskRepository;
 import com.example.demo.repositories.UsersRepository;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
@@ -15,26 +17,37 @@ import java.util.Objects;
 
 @Service
 public class TaskService {
-
-
     private final TaskRepository taskRepository;
     private final UsersRepository usersRepository;
 
     public TaskService(TaskRepository taskRepository, UsersRepository usersRepository) {
         this.taskRepository = taskRepository;
         this.usersRepository = usersRepository;
+
+
     }
 
+public User getAuth(){
+    Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+    String username = auth.getName();
+    User currentUser= usersRepository.findByName(username);
+    if (currentUser == null) {
+        throw new org.springframework.security.authentication.BadCredentialsException("Not Authorized");
 
+    }
+    return  currentUser;
+}
 
-    public List<Task> getUsersTasks(Long userId) {
-
-        return taskRepository.findByUserId(userId);
+    public List<Task> getUsersTasks() {
+        User currentUser = getAuth();
+        return taskRepository.findByUserId(currentUser.getId());
     }
 
-    public Object addNewTask(Long id , TaskDto taskDto ) {
-        User user = usersRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+    public Object addNewTask( TaskDto taskDto ) {
+        User currentUser = getAuth();
+        if (currentUser == null) {
+            throw new RuntimeException("User not found");
+        }
         if(taskDto.getDueDate().isBefore(LocalDate.now())){
              throw new RuntimeException("Due date cannot be in the past");
         }
@@ -43,20 +56,35 @@ public class TaskService {
                 taskDto.getDescription(),
                 taskDto.getStatus(),
                 taskDto.getDueDate(),
-                user
+                currentUser
         );
-        Object savedTask;
-        savedTask = taskRepository.save(task);
-        return  savedTask;
-    }
+        return taskRepository.save(task);
 
-    public boolean deleteTask(Long id  ) {
+    }
+    public boolean deleteTask(Long id) {
+        User currentUser = getAuth();
+        Task task = taskRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Task not found"));
+        if (!task.getUser().getId().equals(currentUser.getId())) {
+            throw new org.springframework.security.access.AccessDeniedException("You cannot delete this task");
+        }
           taskRepository.deleteById(id);
           return  true;
     }
-    public Task editTask(Long id , TaskDto newTask){
-        Task  oldTask=taskRepository.findById(id).get();
 
+    public Task editTask(Long id , TaskDto newTask){
+        User currentUser = getAuth();
+
+        if (currentUser == null) {
+            throw new org.springframework.security.authentication.BadCredentialsException("Not Authorized");
+        }
+
+        Task oldTask = taskRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Task not found"));
+
+        if (!oldTask.getUser().getId().equals(currentUser.getId())) {
+            throw new org.springframework.security.access.AccessDeniedException("You cannot edit this task");
+        }
         if (Objects.nonNull(newTask.getTitle()) && !"".equalsIgnoreCase(newTask.getTitle())) {
             oldTask.setTitle(newTask.getTitle());
         }
@@ -67,9 +95,10 @@ public class TaskService {
             oldTask.setStatus(newTask.getStatus());
         }
         if (Objects.nonNull(newTask.getDueDate()) && newTask.getDueDate() instanceof LocalDate ) {
-            oldTask.setStatus(newTask.getStatus());
+            oldTask.setDueDate(newTask.getDueDate());
         }
 
 return taskRepository.save(oldTask);
     }
+
 }
